+++ src/functionalTest/kotlin/com/autonomousapps/fixtures/RootProject.kt
@@ -6,8 +6,11 @@ import java.io.File
  * Typical root project of an Android build. Contains a `settings.gradle` and `build.gradle`. [agpVersion] will be null
  * for a [MultiModuleJavaLibraryProject].
  */
-class RootProject(librarySpecs: List<LibrarySpec>? = null, agpVersion: String? = null)
-    : RootGradleProject(File(WORKSPACE)) {
+class RootProject(
+    librarySpecs: List<LibrarySpec>? = null,
+    agpVersion: String? = null,
+    extensionSpec: String = ""
+) : RootGradleProject(File(WORKSPACE)) {
 
     override val variant: String? = null
 
@@ -26,25 +29,27 @@ class RootProject(librarySpecs: List<LibrarySpec>? = null, agpVersion: String? =
         """.trimMargin("|"))
 
         withBuildFile("""
-            buildscript {
-                repositories {
-                    google()
-                    jcenter()
-                }
-                dependencies {
-                    ${agpVersion?.let { "classpath 'com.android.tools.build:gradle:$it'" } ?: ""}
-                    classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.61"
-                }
-            }
-            plugins {
-                id('com.autonomousapps.dependency-analysis')
-            }
-            subprojects {
-                repositories {
-                    google()
-                    jcenter()
-                }
-            }
+            |buildscript {
+            |    repositories {
+            |        google()
+            |        jcenter()
+            |    }
+            |    dependencies {
+            |        ${agpVersion?.let { "classpath 'com.android.tools.build:gradle:$it'" } ?: ""}
+            |        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.61"
+            |    }
+            |}
+            |plugins {
+            |    id('com.autonomousapps.dependency-analysis')
+            |}
+            |subprojects {
+            |    repositories {
+            |        google()
+            |        jcenter()
+            |    }
+            |}
+            |
+            |$extensionSpec
         """)
     }
 }