+++ src/main/kotlin/com/autonomousapps/internal/Advisor.kt
@@ -1,6 +1,11 @@
 package com.autonomousapps.internal
 
+import com.autonomousapps.Behavior
+import com.autonomousapps.Ignore
+import com.autonomousapps.Warn
 import org.gradle.api.GradleException
+
+// TODO add a compute() function, and then simplify the getXXX() functions.
 
 /**
  * Produces human- and machine-readable "advice" for how to modify a project's build scripts for a healthy build.
@@ -9,8 +14,13 @@ internal class Advisor(
     private val unusedDirectComponents: List<UnusedDirectComponent>,
     private val usedTransitiveComponents: List<TransitiveComponent>,
     private val abiDeps: List<Dependency>,
-    private val allDeclaredDeps: List<Dependency>
+    private val allDeclaredDeps: List<Dependency>,
+    private val ignoreSpec: IgnoreSpec
 ) {
+
+    internal val filterRemove = ignoreSpec.filterRemove
+    internal val filterAdd = ignoreSpec.filterAdd
+    internal val filterChange = ignoreSpec.filterChange
 
     private val advices = mutableSetOf<Advice>()
 
@@ -24,11 +34,19 @@ internal class Advisor(
         }.toSortedSet()
     }
 
+    /**
+     * Should add used transitive dependencies.
+     */
     fun getAddAdvice(): String? {
         val undeclaredApiDeps = abiDeps.filter { it.configurationName == null }
+            // Exclude any which are to be ignored per user configuration
+            .filterNot { ignoreSpec.shouldIgnoreUsedTransitiveDep(it) }
+
         val undeclaredImplDeps = usedTransitiveComponents.map { it.dependency }
             // Exclude any transitives which will be api dependencies
             .filterNot { trans -> undeclaredApiDeps.any { api -> api == trans } }
+            // Exclude any which are to be ignored per user configuration
+            .filterNot { ignoreSpec.shouldIgnoreUsedTransitiveDep(it) }
 
         if (undeclaredApiDeps.isEmpty() && undeclaredImplDeps.isEmpty()) {
             return null
@@ -61,12 +79,17 @@ internal class Advisor(
         }
     }
 
+    /**
+     * Should remove unused dependencies.
+     */
     fun getRemoveAdvice(): String? {
         val completelyUnusedDeps = unusedDirectComponents
             // This filter was to help find "completely unused" dependencies. However, I (provisionally) think it best
             // to suggest users remove ALL unused dependencies and replace them with the used-transitives.
             //.filter { it.usedTransitiveDependencies.isEmpty() }
             .map { it.dependency }
+            // Exclude any which are to be ignored per user configuration
+            .filterNot { ignoreSpec.shouldIgnoreUnusedDep(it) }
 
         if (completelyUnusedDeps.isEmpty()) {
             return null
@@ -81,12 +104,18 @@ internal class Advisor(
         }
     }
 
+    /**
+     * Should fix dependencies on the wrong configuration.
+     */
     fun getChangeAdvice(): String? {
         val shouldBeApi = abiDeps
             // Filter out those with a null configuration, as they are transitive. They will be reported in "addAdvice"
             .filterNot { it.configurationName == null }
             // Filter out those with an "api" configuration, as they're already correct.
             .filterNot { it.configurationName!!.contains("api", ignoreCase = true) }
+            // Exclude any which are to be ignored per user configuration
+            .filterNot { ignoreSpec.shouldIgnoreIncorrectConfiguration(it) }
+
         val shouldBeImpl = allDeclaredDeps
             // filter out those that are already a flavor of implementation
             .filterNot { it.configurationName!!.contains("implementation", ignoreCase = true) }
@@ -96,6 +125,8 @@ internal class Advisor(
                     .filter { it.configurationName?.contains("api", ignoreCase = true) == true }
                     .find { api -> api == declared } != null
             }
+            // Exclude any which are to be ignored per user configuration
+            .filterNot { ignoreSpec.shouldIgnoreIncorrectConfiguration(it) }
 
         if (shouldBeApi.isEmpty() && shouldBeImpl.isEmpty()) {
             return null
@@ -132,4 +163,46 @@ internal class Advisor(
         } else {
             "\"${dependency.identifier}:${dependency.resolvedVersion}\""
         }
+
+    internal class IgnoreSpec(
+        private val anyBehavior: Behavior = Warn(),
+        private val unusedDependenciesBehavior: Behavior = Warn(),
+        private val usedTransitivesBehavior: Behavior = Warn(),
+        private val incorrectConfigurationsBehavior: Behavior = Warn()
+    ) {
+        private val shouldIgnoreAll = anyBehavior is Ignore
+        internal val filterRemove = shouldIgnoreAll || unusedDependenciesBehavior is Ignore
+        internal val filterAdd = shouldIgnoreAll || usedTransitivesBehavior is Ignore
+        internal val filterChange = shouldIgnoreAll || incorrectConfigurationsBehavior is Ignore
+
+        internal fun shouldIgnoreUnusedDep(dep: Dependency): Boolean {
+            return if (anyBehavior is Ignore || unusedDependenciesBehavior is Ignore) {
+                // If we're just ignoring everything, ignore everything
+                true
+            } else {
+                // Otherwise, ignore if it's been specifically configured to be ignored
+                anyBehavior.filter.plus(unusedDependenciesBehavior.filter).contains(dep.identifier)
+            }
+        }
+
+        internal fun shouldIgnoreUsedTransitiveDep(dep: Dependency): Boolean {
+            return if (anyBehavior is Ignore || usedTransitivesBehavior is Ignore) {
+                // If we're just ignoring everything, ignore everything
+                true
+            } else {
+                // Otherwise, ignore if it's been specifically configured to be ignored
+                anyBehavior.filter.plus(usedTransitivesBehavior.filter).contains(dep.identifier)
+            }
+        }
+
+        internal fun shouldIgnoreIncorrectConfiguration(dep: Dependency): Boolean {
+            return if (anyBehavior is Ignore || incorrectConfigurationsBehavior is Ignore) {
+                // If we're just ignoring everything, ignore everything
+                true
+            } else {
+                // Otherwise, ignore if it's been specifically configured to be ignored
+                anyBehavior.filter.plus(incorrectConfigurationsBehavior.filter).contains(dep.identifier)
+            }
+        }
+    }
 }