+++ src/main/kotlin/com/autonomousapps/tasks/AdviceAggregateReportTask.kt
@@ -55,7 +55,9 @@ abstract class AdviceAggregateReportTask : DefaultTask() {
         projectReportFile.writeText(adviceReports.toJson())
         projectReportPrettyFile.writeText(adviceReports.toPrettyString())
 
-        logger.quiet("Advice report (aggregated) : ${projectReportFile.path}")
-        logger.quiet("(pretty-printed)           : ${projectReportPrettyFile.path}")
+        if (adviceReports.isNotEmpty()) {
+            logger.quiet("Advice report (aggregated) : ${projectReportFile.path}")
+            logger.quiet("(pretty-printed)           : ${projectReportPrettyFile.path}")
+        }
     }
 }